refactor: ProductCard Props 위치 utils->types 이동
