+++ src/shared/types/productCard.ts
